Make ORTEC's dynamic params available from Python
I am unsure if we should use these, but we should test before discarding the idea.

+++ hgs_vrptw/include/Matrix.h
@@ -10,7 +10,7 @@
 // element
 class Matrix
 {
-    int cols_;               // The number of columns of the matrix
+    size_t cols_;            // The number of columns of the matrix
     std::vector<int> data_;  // The vector where all the data is stored (this
                              // represents the matrix)
 
@@ -19,18 +19,18 @@ public:
     Matrix() : cols_(0), data_(0) {}
 
     // Constructor: create a matrix of size dimension by dimension, using a C++
-    // vector of size dimension * dimension
-    explicit Matrix(int dimension)
+    // vector of size dimension * dimension.
+    explicit Matrix(size_t dimension)
         : cols_(dimension), data_(dimension * dimension)
     {
     }
 
-    [[nodiscard]] inline int &operator()(int row, int col)  // for modification
+    [[nodiscard]] inline int &operator()(size_t row, size_t col)
     {
         return data_[cols_ * row + col];
     }
 
-    [[nodiscard]] inline int operator()(int row, int col) const  // for access
+    [[nodiscard]] inline int operator()(size_t row, size_t col) const
     {
         return data_[cols_ * row + col];
     }
